ViewModelTestLanguage: Refactor conditional cell style handling
Note: It seems that later hiding/re-showing cells is not supported directly by MPS. Only the "show if" can be used, but it leads to not adding the cell at all to the editor-collections.
Therefore, we will not use a conditional cell hiding for multiple shown AST parts for the first time.